validatePurchase: Make sure the return value is limited to 4 decimal digits.

+++ classes/com/sapienter/jbilling/server/user/UserBL.java
@@ -1304,7 +1304,7 @@ public class UserBL extends ResultList
         }
     }
 
-    public Double validatePurchase(BigDecimal amount) {
+    public BigDecimal validatePurchase(BigDecimal amount) {
         if (user.getCustomer() == null) {
             return null;
         }
